Allow plotting multiple audio callback analyses
This will differentiate audio callbacks by their marker name.
A given identifying name 'XX' needs to have markers
'XX real-time budget' and 'XX::DataCallback' for this to work as
intended.

Useful when combining for instance an AudioCallbackDriver and an
AudioInputSource for a multi-mic test case.

Note that this is also backwards compatible with marker names
"Real-time budget" and "AudioCallbackDriver::DataCallback".